requirements.txt
flet>=0.21.0
openai>=1.0.0
python-dotenv>=1.0.0
deep-translator>=1.11.4